Revert "Preprocess /overview response (#1958)"
This reverts commit 6571b10a084e85c439e49aec84ab880cd666f8c2.
