+++ defi/serverless.yml
@@ -393,12 +393,6 @@ functions:
     timeout: 900
     events:
       - schedule: cron(0 * * * ? *)
-  processProtocolsSummary:
-    handler: src/adaptors/handlers/processProtocolsSummary.default
-      timeout: 60
-    environment:
-      VOLUMES_WEBHOOK: ${file(./env.js):VOLUMES_WEBHOOK}
-      FEES_WEBHOOK: ${file(./env.js):FEES_WEBHOOK}
   getOverview:
     handler: src/adaptors/handlers/getOverview/index.default
     timeout: 30
