+++ defi/src/adaptors/handlers/storeAdaptorData/helpers.ts
@@ -31,11 +31,10 @@ export const STORE_ERROR = "STORE_ERROR"
 export function processRejectedPromises(rejectedResults: IRunAdapterResponseRejected[], rawRecord: RawRecordMap, dexName: string, ATTRIBUTE_KEYS: IJSON<string>) {
     for (const [RECORD_TYPE, ATTRIBUTE] of Object.entries(ATTRIBUTE_KEYS)) {
         for (const result of rejectedResults) {
-            console.error(`${STORE_ERROR}:${dexName}:Rejected:${ATTRIBUTE}: \nTIMESTAMP: ${result.timestamp}`, result)
+            console.error(`${STORE_ERROR}:${dexName}:Rejected: \nTIMESTAMP: ${result.timestamp}`, result)
             if (!rawRecord[RECORD_TYPE]) rawRecord[RECORD_TYPE] = {}
             const recordChain = rawRecord[RECORD_TYPE][result.chain]
             if (typeof recordChain === 'number') return
-            // TODO: change errors to another column
             rawRecord[RECORD_TYPE] = {
                 ...rawRecord[RECORD_TYPE],
                 [result.chain]: {
