--- defi/src/adaptors/utils/storeR2Response/index.ts
@@ -1,20 +1,0 @@
-import { getR2, storeR2 } from "../../../utils/r2";
-import { IGetOverviewResponseBody } from "../../handlers/processProtocolsSummary";
-
-const ADAPTORS_FOLDER_KEY = 'dimensions'
-
-export const cacheResponseOnR2 = async (key: string, response: string) => {
-    await storeR2(`${ADAPTORS_FOLDER_KEY}/${key}`, response)
-}
-
-export const getCachedResponseOnR2 = async (key: string): Promise<IGetOverviewResponseBody | undefined> => {
-    const objectString = await getR2(`${ADAPTORS_FOLDER_KEY}/${key}`)
-    if (objectString) {
-        try {
-            return JSON.parse(objectString) as IGetOverviewResponseBody
-        } catch (error) {
-            console.error(error)
-        }
-    }
-    return undefined
-}
